% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataformat}
\alias{dataformat}
\title{Data formats}
\usage{
dataformat(file_name, check_exists = TRUE)
}
\arguments{
\item{file_name}{Path to a file.}

\item{check_exists}{Check if file exists.}
}
\value{
An object of class \code{dataformat}
}
\description{
Functions to help selecting the correct loader for a given file name.
}
\examples{
\dontrun{dataformat("my/data/file.csv.bz2")}
}

