% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{package}
\name{LoadMyData}
\alias{LoadMyData}
\alias{LoadMyData-package}
\alias{package-LoadMyData}
\title{Automates the creation of new statistical analysis projects.}
\description{
LoadMyData provides functions to automatically build a directory
structure for a new R project. Using this structure, LoadMyData
automates data loading, preprocessing, library importing and unit
testing.
}
\examples{
\dontrun{create.project('project_name')

setwd('project_name')
load.project()}
}
\references{
This code is inspired by the skeleton structure used by
  Ruby on Rails.
}

