% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.bz2}
\alias{reader.dataformat.bz2}
\title{Read a bzip2-compressed remote file}
\usage{
\method{reader}{dataformat.bz2}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments passed to the reader invoked.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load data from a bzip2-compressed file in the format
specified by the "parent" extension, i.e. the extension that remains after
stripping the \code{.bz2} extension.
}

