% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.csv.R
\name{reader.dataformat.csv}
\alias{reader.dataformat.csv}
\title{Read a comma separated values (.csv) file.}
\usage{
\method{reader}{dataformat.csv}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the CSV file format into
the specified global variable binding.
}
\examples{
\dontrun{reader.csv('example.csv', 'data/example.csv', 'example')}
}

