% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.dbf.R
\name{reader.dataformat.dbf}
\alias{reader.dataformat.dbf}
\title{Read an XBASE file with a .dbf file extension.}
\usage{
\method{reader}{dataformat.dbf}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load all of the data sets stored in the specified
XBASE file into the global environment.
}
\examples{
\dontrun{reader.dbf('example.dbf', 'data/example.dbf', 'example')}
}

