% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.dta.R
\name{reader.dataformat.dta}
\alias{reader.dataformat.dta}
\title{Read a Stata file with a .dta file extension.}
\usage{
\method{reader}{dataformat.dta}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Stata file into memory.
}
\examples{
\dontrun{reader.stata('example.stata', 'data/example.stata', 'example')}
}

