% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.epiinfo.R
\name{reader.dataformat.epiinfo}
\alias{reader.dataformat.epiinfo}
\title{Read an Epi Info file with a .rec file extension.}
\usage{
\method{reader}{dataformat.epiinfo}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load all of the data sets stored in the specified
Epi Info file into the global environment.
}
\examples{
\dontrun{reader.epiinfo('example.rec', 'data/example.rec', 'example')}
}

