% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.gz.R
\name{reader.dataformat.gz}
\alias{reader.dataformat.gz}
\title{Read a gzip-compressed remote file}
\usage{
\method{reader}{dataformat.gz}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments passed to the reader invoked.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load data from a gzip-compressed file in the format
specified by the "parent" extension, i.e. the extension that remains after
stripping the \code{.gz} extension.
}

