% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.mp3}
\alias{reader.dataformat.mp3}
\title{Read an MP3 file with a .mp3 file extension.}
\usage{
\method{reader}{dataformat.mp3}(filename, data.file, variable.name)
}
\arguments{
\item{filename}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified MP3 file into memory using the
tuneR package. This is useful for working with music files as a data
set.
}
\examples{
\dontrun{reader.mp3('example.mp3', 'data/example.mp3', 'example')}
}

