% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.mtp}
\alias{reader.dataformat.mtp}
\title{Read a Minitab Portable Worksheet with an .mtp3 file extension.}
\usage{
\method{reader}{dataformat.mtp}(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Minitab Portable Worksheet into
memory.
}
\examples{
\dontrun{reader.mtp('example.mtp', 'data/example.mtp', 'example')}
}

