% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.octave.R
\name{reader.dataformat.octave}
\alias{reader.dataformat.octave}
\title{Read an Octave file with a .m file extension.}
\usage{
\method{reader}{dataformat.octave}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Octave file into memory.
}
\examples{
\dontrun{reader.octave('example.m', 'data/example.m', 'example')}
}

