% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.ppm.R
\name{reader.dataformat.ppm}
\alias{reader.dataformat.ppm}
\title{Read a PPM file with a .ppm file extension.}
\usage{
\method{reader}{dataformat.ppm}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified PPM file into memory using the
pixamp package. This is useful for working with image files as a data
set.
}
\examples{
\dontrun{reader.ppm('example.ppm', 'data/example.ppm', 'example')}
}

