% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.r.R
\name{reader.dataformat.r}
\alias{reader.dataformat.R}
\alias{reader.dataformat.r}
\title{Read an R source file with a .R file extension.}
\usage{
\method{reader}{dataformat.r}(x, ...)

\method{reader}{dataformat.R}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will call source on the specified R file, executing the
code inside of it as a way of generating data sets dynamically, as in
many Monte Carlo applications.
}

