% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.RData}
\alias{reader.dataformat.RData}
\alias{reader.dataformat.rda}
\title{Read an RData file with a .rdata or .rda file extension.}
\usage{
\method{reader}{dataformat.RData}(x, data.file, variable.name, ...)

\method{reader}{dataformat.rda}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified RData file into memory using the
\code{\link{load}} function. This may generate many data sets
simultaneously.
}

