% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.sav.R
\name{reader.dataformat.sav}
\alias{reader.dataformat.sav}
\title{Read an SPSS file with a .sav file extension.}
\usage{
\method{reader}{dataformat.sav}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified SPSS file into memory. It will
convert the resulting list object into a data frame before inserting the
data set into the global environment.
}

