% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.systat.R
\name{reader.dataformat.systat}
\alias{reader.dataformat.systat}
\title{Read a Systat file with a .sys or .syd file extension.}
\usage{
\method{reader}{dataformat.systat}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Systat file into memory.
}
\examples{
\dontrun{reader.systat('example.sys', 'data/example.sys', 'example')}
}

