% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.tsv.R
\name{reader.dataformat.tsv}
\alias{reader.dataformat.tab}
\alias{reader.dataformat.tsv}
\title{Read a tab separated values (.tsv or .tab) file.}
\usage{
\method{reader}{dataformat.tsv}(x, ...)

\method{reader}{dataformat.tab}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the TSV file format into
the specified global variable binding.
}

