% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.xls.R
\name{reader.dataformat.xls}
\alias{reader.dataformat.xls}
\title{Read an Excel 2004 file with a .xls file extension.}
\usage{
\method{reader}{dataformat.xls}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Excel file into memory using the
gdata package. Each sheet of the Excel workbook will be read into a
separate variable in the global environment.
}
\examples{
\dontrun{reader.xls('example.xls', 'data/example.xls', 'example')}
}

