% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.xlsx.R
\name{reader.dataformat.xlsx}
\alias{reader.dataformat.xlsx}
\title{Read an Excel 2007 file with a .xlsx file extension.}
\usage{
\method{reader}{dataformat.xlsx}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified Excel file into memory using the
xlsx package. Each sheet of the Excel workbook will be read into a
separate variable in the global environment.
}
\examples{
\dontrun{reader.xlsx('example.xlsx', 'data/example.xlsx', 'example')}
}

