% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.xport.R
\name{reader.dataformat.xport}
\alias{reader.dataformat.xport}
\alias{reader.dataformat.xpt}
\title{Read an SAS XPort file with a .xpt or .xport file extension.}
\usage{
\method{reader}{dataformat.xport}(x, ...)

\method{reader}{dataformat.xpt}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified SAS XPort file into memory.
}

