% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.R, R/reader.arff.R, R/reader.csv.R, R/reader.csv2.R, R/reader.db.R, R/reader.dbf.R, R/reader.dta.R, R/reader.epiinfo.R, R/reader.mp3.R, R/reader.mtp.R, R/reader.octave.R, R/reader.ppm.R, R/reader.sav.R, R/reader.systat.R, R/reader.tsv.R, R/reader.wsv.R, R/reader.xls.R, R/reader.xlsx.R, R/reader.xport.R
\name{reader}
\alias{reader}
\alias{reader.dataformat.arff}
\alias{reader.dataformat.csv}
\alias{reader.dataformat.csv2}
\alias{reader.dataformat.db}
\alias{reader.dataformat.dbf}
\alias{reader.dataformat.dta}
\alias{reader.dataformat.epiinfo}
\alias{reader.dataformat.mp3}
\alias{reader.dataformat.mtp}
\alias{reader.dataformat.octave}
\alias{reader.dataformat.ppm}
\alias{reader.dataformat.sav}
\alias{reader.dataformat.systat}
\alias{reader.dataformat.tab}
\alias{reader.dataformat.tsv}
\alias{reader.dataformat.txt}
\alias{reader.dataformat.wsv}
\alias{reader.dataformat.xls}
\alias{reader.dataformat.xlsx}
\alias{reader.dataformat.xport}
\alias{reader.dataformat.xpt}
\title{Read a data file}
\usage{
reader(x, ...)
}
\arguments{
\item{x}{An object of class \code{dataformat}, will be coerced using
\code{\link{as.dataformat}} if necessary.}

\item{...}{Further arguments.}
}
\value{
A list where each element corresponds to a loaded data item.
  This is always a named list; the name is derived from the input file name
  using \code{\link{clean.variable.name}}.  Readers that support files with
  multiple data sets (e.g., for SQL databases, Excel spreadsheets,
  ZIP archives, \ldots) will return a nested list.
}
\description{
This function loads a data set.  The data format is determined by the
file extension, or by the class if an object of class \code{\link{dataformat}}
is passed.
}
\details{
The following file extensions are supported:

\code{arff}: Weka Attribute-Relation File Format via \code{foreign::\link[foreign]{read.arff}}

\code{csv}: CSV file (comma-separated) via \code{\link{read.csv}}

\code{csv2}: CSV file (semicolon-separated) via \code{\link{read.csv}(sep = ';')}

\code{xlsx}: SQlite3 database via  \code{RSQLite::\link[RSQLite]{dbConnect}}.
  Each table will be read, the return value is a list. (If you want to specify
  a single table or query to execute against the database, move it elsewhere
  and use a \code{.sql} file interpreted by \code{\link{reader.dataformat.sql}}.)

\code{dbf}: xBase file via \code{foreign::\link[foreign]{read.dbf}}

\code{dta}: Stata file via \code{foreign::\link[foreign]{read.dta}}

\code{rec}: Epi Info file via \code{foreign::\link[foreign]{read.epiinfo}}

\code{mp3}: MP3 file format via \code{tuneR::\link[tuneR]{readMP3}}.
  Useful for working with audio files as a data set.

\code{mtp}: Minitab Portable Worksheet via \code{foreign::\link[foreign]{read.mtp}}

\code{m}: Octave format via \code{foreign::\link[foreign]{read.octave}}

\code{ppm}: PPM file format via \code{pixmap::\link[pixmap]{read.pnm}}.
  Useful for working with image files as a data set.

\code{sav}: SPSS data files via \code{foreign::\link[foreign]{read.spss}}

\code{syd}, \code{sys}: Systat file via \code{foreign::\link[foreign]{read.systat}}

\code{tab}, \code{tsv}: TSV file (tab-separated) via \code{\link{read.csv}(sep = '\\t')}

\code{txt}, \code{wsv}: WSV file (whitespace-separated) via \code{\link{read.table}}

\code{xls}: Excel 2004 files via \code{gdata::\link[gdata]{read.xls}}.
  Each sheet will be read, the return value is a list.

\code{xlsx}: Excel 2007 files via \code{xlsx::\link[xlsx]{read.xlsx}}.
  Each sheet will be read, the return value is a list.

\code{xport}, \code{xpt}: SAS Xport file via \code{foreign::\link[foreign]{read.xport}}
}
\examples{
\dontrun{reader('example.csv')}
}

