% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.spss}
\alias{reader.spss}
\title{Read an SPSS file with a .sav file extension.}
\usage{
reader.spss(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified SPSS file into memory. It will
convert the resulting list object into a data frame before inserting the
data set into the global environment.
}
\examples{
\dontrun{reader.spss('example.sav', 'data/example.sav', 'example')}
}

