% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/dataformat.R
\name{use_extension}
\alias{explicit_extension}
\alias{parent_extension}
\alias{use_extension}
\title{Placeholders for override_extension}
\usage{
use_extension()

parent_extension()

explicit_extension(extension = NULL, objname = NULL)
}
\arguments{
\item{extension}{Alternate extension}
}
\description{
Use as values for the \code{override_extension} parameter to
\code{\link{dataformat}}.
}

