% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dataformat}
\alias{as.dataformat}
\alias{dataformat}
\alias{is.dataformat}
\title{Data formats}
\usage{
dataformat(file_name, check_exists = TRUE, override_extension = NULL)

as.dataformat(file_name, check_exists = TRUE, override_extension = NULL)

is.dataformat(x)
}
\arguments{
\item{file_name}{Path to a file.}

\item{check_exists}{Check if file exists (default).}

\item{override_extension}{Alternative file extension (default: use extension
from \code{file_name}).}

\item{x}{An object.}
}
\value{
An object of class \code{dataformat}
}
\description{
Functions to help selecting the correct loader for a given file name.
}
\examples{
\dontrun{dataformat("my/data/file.csv.bz2")}
}

