% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/dataformat.R
\name{dataformat}
\alias{as.dataformat}
\alias{dataformat}
\alias{is.dataformat}
\title{Data formats}
\usage{
dataformat(x, override_extension = NULL, check_exists = is.character(x),
  ...)

as.dataformat(x, override_extension = NULL, check_exists = is.character(x),
  ...)

is.dataformat(x)
}
\arguments{
\item{x}{Path to a file.}

\item{override_extension}{Alternative file extension (default: use extension
from \code{x}).}

\item{check_exists}{Check if file exists (default).}

\item{...}{Further parameters.}
}
\value{
An object of class \code{dataformat}
}
\description{
Functions to help selecting the correct reader for a given file name.
}
\examples{
\dontrun{dataformat("my/data/file.csv.bz2")}
}

