% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.csv}
\alias{reader.dataformat.csv}
\title{Read a comma separated values (.csv) file.}
\usage{
\method{reader}{dataformat.csv}(x, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}

\item{data.file}{The name of the data file to be read.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the CSV file format into
the specified global variable binding.
}
\examples{
\dontrun{reader.csv('example.csv', 'data/example.csv', 'example')}
}

