% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.db.R
\name{reader.dataformat.db}
\alias{reader.dataformat.db}
\title{Read a SQlite3 database with a (.db) file extension.}
\usage{
\method{reader}{dataformat.db}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load all of the data sets stored in the SQlite3
database into the global environment. If you want to specify a single
table or query to execute against the database, move it elsewhere and
use a .sql file interpreted by \code{\link{reader.dataformat.sql}}.
}
\examples{
\dontrun{reader.db('example.db', 'data/example.db', 'example')}
}

