% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.mp3.R
\name{reader.dataformat.mp3}
\alias{reader.dataformat.mp3}
\title{Read an MP3 file with a .mp3 file extension.}
\usage{
\method{reader}{dataformat.mp3}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified MP3 file into memory using the
tuneR package. This is useful for working with music files as a data
set.
}
\examples{
\dontrun{reader.mp3('example.mp3', 'data/example.mp3', 'example')}
}

