% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.ppm}
\alias{reader.dataformat.ppm}
\title{Read a PPM file with a .ppm file extension.}
\usage{
\method{reader}{dataformat.ppm}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified PPM file into memory using the
pixamp package. This is useful for working with image files as a data
set.
}
\examples{
\dontrun{reader.ppm('example.ppm', 'data/example.ppm', 'example')}
}

