% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.r}
\alias{reader.dataformat.r}
\title{Read an R source file with a .R file extension.}
\usage{
\method{reader}{dataformat.r}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will call source on the specified R file, executing the
code inside of it as a way of generating data sets dynamically, as in
many Monte Carlo applications.
}
\examples{
\dontrun{reader.r('example.R', 'data/example.R', 'example')}
}

