% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.rdata}
\alias{reader.dataformat.rdata}
\title{Read an RData file with a .rdata or .rda file extension.}
\usage{
\method{reader}{dataformat.rdata}(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified RData file into memory using the
\code{\link{load}} function. This may generate many data sets
simultaneously.
}
\examples{
\dontrun{reader.rdata('example.RData', 'data/example.RData', 'example')}
}

