% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.sav}
\alias{reader.dataformat.sav}
\title{Read an SPSS file with a .sav file extension.}
\usage{
\method{reader}{dataformat.sav}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified SPSS file into memory. It will
convert the resulting list object into a data frame before inserting the
data set into the global environment.
}

