% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.tsv}
\alias{reader.dataformat.tsv}
\title{Read a tab separated values (.tsv or .tab) file.}
\usage{
\method{reader}{dataformat.tsv}(x, data.file, variable.name, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}

\item{...}{Further arguments.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load a data set stored in the TSV file format into
the specified global variable binding.
}
\examples{
\dontrun{reader.tsv('example.tsv', 'data/example.tsv', 'example')}
}

