% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader.dataformat.xport}
\alias{reader.dataformat.xport}
\title{Read an XPort file with a .xport file extension.}
\usage{
\method{reader}{dataformat.xport}(filename, data.file, variable.name)
}
\arguments{
\item{filename}{The path to the data set to be loaded.}

\item{data.file}{The name of the data file to be read.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load the specified XPort file into memory.
}
\examples{
\dontrun{reader.xport('example.xport', 'data/example.xport', 'example')}
}

