% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.zip.R
\name{reader.dataformat.zip}
\alias{reader.dataformat.zip}
\title{Read a zip-compressed remote file}
\usage{
\method{reader}{dataformat.zip}(x, ...)
}
\arguments{
\item{x}{The path to the data set to be loaded.}

\item{...}{Further arguments passed to the reader invoked.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will load data from a zip-compressed file.  Only the first file
will be extracted.
}

