% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reader}
\alias{reader}
\title{Read a data file}
\usage{
reader(x, ...)
}
\arguments{
\item{x}{An object of class \code{dataformat}, will be coerced using
\code{\link{as.dataformat}} if necessary.}

\item{...}{Further parameters}
}
\value{
A list where each element corresponds to a loaded data item.
  Often, this will be an unnamed list with just one element.
}
\description{
This function loads a data set.  The data format is determined by the
file extension, or by the class if an object of class \code{\link{dataformat}}
is passed.
}
\examples{
\dontrun{reader('example.csv')}
}

