% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.R, R/reader.arff.R, R/reader.octave.R, R/reader.sav.R
\name{reader}
\alias{reader}
\alias{reader.dataformat.arff}
\alias{reader.dataformat.octave}
\alias{reader.dataformat.sav}
\title{Read a data file}
\usage{
reader(x, ...)
}
\arguments{
\item{x}{An object of class \code{dataformat}, will be coerced using
\code{\link{as.dataformat}} if necessary.}

\item{...}{Further arguments.}
}
\value{
A list where each element corresponds to a loaded data item.
  This is always a named list; the name is derived from the input file name
  using \code{\link{clean.variable.name}}.  Readers that support files with
  multiple data sets (e.g., for SQL databases, Excel spreadsheets,
  ZIP archives, \ldots) will return a nested list.
}
\description{
This function loads a data set.  The data format is determined by the
file extension, or by the class if an object of class \code{\link{dataformat}}
is passed.
}
\details{
The following file extensions are supported:

\code{arff}: Weka Attribute-Relation File Format via \code{foreign::\link[foreign]{read.arff}}

\code{m}: Octave format via \code{foreign::\link[foreign]{read.octave}}

\code{sav}: SPSS data files via \code{foreign::\link[foreign]{read.spss}}
}
\examples{
\dontrun{reader('example.csv')}
}

