% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/reader.R, R/reader.arff.R, R/reader.csv.R, R/reader.dbf.R, R/reader.epiinfo.R, R/reader.mp3.R, R/reader.mtp.R, R/reader.octave.R, R/reader.ppm.R, R/reader.sav.R
\name{reader}
\alias{reader}
\alias{reader.dataformat.arff}
\alias{reader.dataformat.csv}
\alias{reader.dataformat.dbf}
\alias{reader.dataformat.epiinfo}
\alias{reader.dataformat.mp3}
\alias{reader.dataformat.mtp}
\alias{reader.dataformat.octave}
\alias{reader.dataformat.ppm}
\alias{reader.dataformat.sav}
\title{Read a data file}
\usage{
reader(x, ...)
}
\arguments{
\item{x}{An object of class \code{dataformat}, will be coerced using
\code{\link{as.dataformat}} if necessary.}

\item{...}{Further arguments.}
}
\value{
A list where each element corresponds to a loaded data item.
  This is always a named list; the name is derived from the input file name
  using \code{\link{clean.variable.name}}.  Readers that support files with
  multiple data sets (e.g., for SQL databases, Excel spreadsheets,
  ZIP archives, \ldots) will return a nested list.
}
\description{
This function loads a data set.  The data format is determined by the
file extension, or by the class if an object of class \code{\link{dataformat}}
is passed.
}
\details{
The following file extensions are supported:

\code{arff}: Weka Attribute-Relation File Format via \code{foreign::\link[foreign]{read.arff}}

\code{dbf}: CSV file (comma-separated) via \code{\link{read.csv}}

\code{dbf}: xBase file via \code{foreign::\link[foreign]{read.dbf}}

\code{rec}: Epi Info file via \code{foreign::\link[foreign]{read.epiinfo}}

\code{mp3}: MP3 file format via \code{tuneR::\link[tuneR]{read.MP3}}.
  Useful for working with audio files as a data set.

\code{mtp3}: Minitab Portable Worksheet via \code{foreign::\link[foreign]{read.mtp}}

\code{m}: Octave format via \code{foreign::\link[foreign]{read.octave}}

\code{ppm}: PPM file format via \code{pixmap::\link[pixmap]{read.pnm}}.
  Useful for working with image files as a data set.

\code{sav}: SPSS data files via \code{foreign::\link[foreign]{read.spss}}
}
\examples{
\dontrun{reader('example.csv')}
}

