% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{use_extension}
\alias{parent_extension}
\alias{use_extension}
\title{Placeholders for override_extension}
\usage{
use_extension()

parent_extension()
}
\description{
Use as values for the \code{override_extension} parameter to
\code{\link{dataformat}}.
}

